Enable testing support for Firefox
This adds a dependency for karma: karma-firefox-launcher, which is essentially Firefox.

It defines a custom broswer FirefoxHeadless that essentially uses the Firefox browser without opening it

Then in angular.json it defines test-headless-firefox which will test the angular application with firefox.

Finally, the test command has been added to angular-ci.yaml.